Changed my Tech Stack.
My previous tech-stack was RAMB (React, Aiohttp, Mysql, Bootstrap) is
very complex for my small brain, I am migrating this to BAHM (Bootstrap,
Aiohttp, Htmx, Mysql) which is much easier for me. This solves a mega
problem RAMB introduces, learning react and its ecosystem. Htmx is a
much easier replacement for React as it abstracts a lot of javascript
code to an extent I can majorly concentrate on my backend. With this
simplification comes a new form of API, HTMX requires the server to
respond with HTML for the components required which can be created using
python templating engines like Mako or Jinja which are well intergrated
with Aiohttp framework.