Making all code-practices terminal capable.
From now on, for maintainability and reusability, I will convert all my
code practices to be terminal capable, ie, they take input from the
terminal and use them in their computations. The advantage of this, is
that, in the future, testing this algorithms will not require me to
change the hard coded inputs just to see how they interact with a
different set of inputs.