Solved Another LeetCode Problem.
Unfortunately this problem took me too long such that leetcode did not
count it as a solved problem, but identified my participation.
TODO: Keep time in these practice quizes
The solution was still amazing to come up with. Also I have separated
code that can be reused in other quizes to the header file `helpers.hpp`