New C++ trick involving member pointers.
I remember learning about storing member functions to be passed around
and there signatures that took the form:
	<return_type> (<class_type>::*<var_name>)(<arg_types>...)[const]
eg
		std::string (person::*func)(std::string const &) const = &person::greet;
		person me { "Simon Nganga" };
		(me.*func)("Faith Njeri"); //-> "Hello Faith Njeri, my name is Simon Nganga"

This was life changing, turns out, this can also be applied to member
variables for a class/struct/compound object using the form:
		<member_type> <class_type>::*<var_name>
eg
		std::string person me { "Simon Nganga" };
		std::string person::*value = &person::name;
		std::cout << "Hello " << me.*value << '\n';
		// -> "Hello Simon Nganga\n"
Now this is nothing short but amazing. With this you can abstract away
the member variable being accessed and just use an overloaded operator()
on the accessor type to get the value of the variable. I have
constructed this on the header file "live/c++/mem_value.hpp".
eg
		auto value = snn::member(&person::name);
		person me { "Simon Nganga" };
		std::cout << "Name: " << value(me) << '\n';
		// -> Name: Simon Nganga
		//You can also convert on read.
		std::cout << "Email: " << value.to<hey::email>(me) << '\n';
		// -> Email: simonnganga@gmail.com
		//The core accessor is a struct object _Member_value object
		//and the core reader method is _Member_value::get while the
		//converter is _Mamber_value::to which is templated and expects
		//a Type T such that the member_type is convertible to T

The objects snn::name and hey::email can be found in the file
live/c++/general_/mem_value.cpp which also contains more examples of
this usage.