Read more on Bash.
Today I looked at solving math using bash, from builtin bash operators
like `(())` to tools like `bc`, `let` and `expr`.
And also `expr` works only on integers.
Bash Scoping is amazing, it is the only language I know that supprts
Dynamic Scope Lookup, which means, a functions variable lookup happens
when it is called and not when it is defined.