Fixed a bug in retry and added ex_alias.
The retry command utility tool from last commit had a bug involving the
--last/--LAST/--- flag that was meant to treat the rest of the arguments
as part of the command to be executed. The problem involved shifting
which I forgot, therefore, the command was the option itself instead of
what follows it.

Also, I cannot pass aliases to the retry command so I created a tool to
get the original command that had been aliased.
Example:
	$ alias hello="echo Hey friend? How are you?"
	$ hello
	Hey friend How are you?
	$ retrythis --- hello
	Running command: 'hello'
	bash: hello: command not found
	bash: hello: command not found
	bash: hello: command not found
	$ retrythis --- $(ex_alias hello)
	Running command: 'hello'
	Hey friend? How are you?
The ex_alias command returns an error if an alias of thet name cannot be
found, or the aliased command if a valid alias is passed.