Solved one leetcode problem.
I don't mean to brag but today I solved a leetcode problem using math
and it beats 100% of other users who submitted their solutions in c++.
I feel alive, energetic and free. This is my highest performance so far
and the mathematical part of it was the most fulfilling.
The other day, I looked into number systems, how to convert from one to
the other, the different operations applicable to each and some other
properties that make them valid number systems like the identity
property. With this knowledge at heart, I attempted a leetcode problem
that involved reversing the digits of a given number and also making
sure it doesn't overflow, if it does, return a 0 instead. To cut it
short, I used some number system magic to solve it and it turned out to
be the fastest algorithm there. Cooooool.