One leet code problem and binary search algorithm.
Today I did one leet code problem that was in the easy tier but on
a deeper look, it was more challenging, it involved finding the square
root of a number to the nearest integer down. I had seen the quiz before
and I took some time to analyze what the question asked, and I wondered how
computers calculated square roots. This made me think of the horrific
algorithm I described in the third commit behind, it involves decaying
the number until it satisfies some predicate which is simply an accuracy
range, I thought it was overcomplicated so I used math to create a
function to get the square root of a number using logarithms. When I
opened the quiz today, unfortunately it required that one must not use
builtin exponentiation operators and functions, simply, calculate the
square root from scratch. Well, my math solution failed since it relied
on pow function to get the square root. I wen back to my horrible idea
and worked on it today, took some time, but eventually got something
working to an expected accuracy, wonderful. After a little research I
found that the algorithm is called the binary search, amazing, I created
the binary search by my self. Hoooooray!!!

There are two implementations, one in python
module`live/python/general_/nthroot.py` with a generic
function`nthroot`, I did use function`pow` in it but to calculate
squareroot, assume simply the variable is being multiplied by itself,
the other implementation is in c++ and stored in the ussual place
`live/others/codepractices/sqrt.cpp` which completely complies with the
rules of the puzzle also described in the file.