Read on Bash Internal Variables.
Today I got to understand the impact of bash internal variables like IFS
in how my arrays are interpreted and constructed. Before this, I had
problems creating arrays from command outputs specifically, how to make
each line of the output an element as a whole in the array without
breaking at the spaces in the output, ie use the newline character as a
delimeter when inserting elements into the array but it turns out the
IFS variable does exactly this, uses whatever it's set to as a
delimeter in operations like inserting into arrays and so much more.
Also remember my `ex_alias` command for returning the command aliased
from an alias name, turns out the BASH_ALIASES variable is an
associative array mapping alias-names to aliased-commands and using it
was much easier than parsing the returned output to extract the command.