Read on Bash data streams and some bash operators.
Data streams by field or line can be controlled using the IFS combined
with the read command to get input in a specified state. You can read
delimetered input section by section ie line / field, to variables or an
array, the read command also allows for timeout specification which is
really amazing in terms of creating streams and user event driven
applications. I also read on some new bash operators, ;;, ;&, ;;&, |,
|&, and -v. The first three are mostly associated with the case
statement, the next two are for joining multiple commands to share a
stream of data as each alters it ready for the next while the last is
used to check if a variable exists, like -e for paths, -v for variables.