Solved the first leet code problem of the week
This was interesting given that it felt a little bit underexplained
though, as usual, yah boy beats 100% of other c++ solutions. The problem
involved converting a string to its integer equivalent under some
specific conditions, simply, implement the std::atoi c++ function.
It had alot of edge cases, and I also noticed one very tiny problem that
has being occurring all around my solutions from leetcode based to
personal, when iterating over a range or just looping in a while loop,
we use variables to represent the state of the solution at any given
moment in the loop, the problem is, when is the state representing the
previous, current and next iteration.
For example:
	uint totalValue = 1;
	for (int i = 0; i < 5; i++) {
		std::cout << '[' << i << "] Before: " << totalValue << '\n';
		totalValue *= 10;
		std::cout << '[' << i << "] After: " << totalValue << '\n';
	}
Output:
	0 Before: 1
	0 After 10
	1 Before 10
	1 After 100
	2 Before 100
	3 After 1000
	4 Before 1000
	5 After 10000
In this example, i represents the current state of the iteration, while
totalValue depends on where it is used, in the before statement, it
represents the state of the last iteration and on the after statement,
it represents the state of the next iteration or maybe the current, who
knows. The problem comes to how you use it, if you wanted it to go along
the loop then use it after assigning it and if it is always behind,
assign it as the last statement in the loop and also after exiting the
loop, if totalValue is to be used, remember its state, is it behind,
ahead or current. This bug is hard to trace down, it bothered me a lot
in this solution as well as the int_exp_growing function I implemented
in python when creating the _create_growing_map function. It was NASTY
to even think about.