Read on first third of asyncio chapter 9
Learned of a new tool called wrk that is simple and can be used to
hammer web backends as many times as specified to test its limits and
get some approximate speed of the backend. I read on how aiohttp and
flask speed differs, flask is fast but aiohttp can get 5 times faster.
I also recreated the database tables from chapter 5 to enable this
chapter and I will have to go back to chapter seven last example which
also relied on the database to see the improvements of using thread
pools in asyncio.