Read second third of asyncio chapter 9
Today's read was, messy. I got to test the new jkuat student portal and
it is super slow, tried to debug the issue and found it to be a caching
issue since they are reloading all the content including the CDN
content, bootstrap and jquery packages to say the least which have to be
downloaded and parsed and most probably will make even more requests to
more resources like fonts, all these time piles up to almost a whole
minute which leads to a very poor user experience. This happens again
every time you move to a different webpage in the site due to completely
no caching, `Cache-Control: no-cache, no-store`. Otherwise, asyncio was
amazing and I definately learned something new about aiohttp and this
awesome framework starlette that is ASGI complient, amazing.

I got to test out caching using a simple factorial, and caching does
improve speed and overall performance.