A little of DSA Today!!
So there is this beautiful function called the 'Wave Function Collapse'
that seems to be pretty cool, it generates patterns based on some
predefined patterns and the best part is that the patterns are connected
and amazing. I'm yet to implement this beauty.
Soo there are what we call greedy algorithms that work by choosing the
local best choice without considering the future's or past's decision
impact which doesn't usually result in the best solution, the most
optimal solution is achieved by using dynamic programming. Though it is
not as simple as that since there are techniques to improve greedy
algorithms, like if the overall solution can be made optimal by making
optimal solutions in the subproblems aka local best solution, then a
greedy solution can return the most optimal solution. But DP still
remains the go-to solution for getting the most optimal solution.