Finished chapter 12 asyncio.
Finally, python asyncio chapter 12 down, kapooosh. This was a tough
chapter and it is amazing how the asyncio.queues module is implemented,
the flow, the inheritance and its general beauty in putters and waiters.

The queues themselves were easier to read and reason about as compared
to the mundane 'I already know' screams at the back of my head. It was a boring
topic.