4 design patterns read.
I thought design patterns were these Godish ideas but they are some form
of solutions to design. I've read 4 design patterns and I am on fire.
The visitor works best in graphy/treeish objects that have to be
iterated one by one, and they are all some how different and can only
support some specific set of operations. The template method is where
the base class holds the deep unchanging logic scarttered in different
methods/functions in which the Subclasses can override to specify what
should be done on specific steps in the algorithm.
The strategy, like visitor, is about separating the logic from the
objects. The objects become data blocbs. The difference being the
strategy d.p is about the same kind of data, how you act on it to get
some desired output. The strategy is injected into the object container
which then calls the strategy to align the objects to the desired form
then it can be used. The facade, my favorite, Its about abstracting away
how something is done from start to end in some simple object, an
interface, which has methods that hit the right spots in a complex
system to get the output desired without the client knowing about the
specifics. We can say, Object to Adoptor as System to Facade.