Read on a few algorithms. DP, Greedy, ...
Just finished 'Grokking Algorithms' chapter 9 from 7. It was about
greedy algorithms, dynamic programming and NP-complete problems.
Some problems cannot be computed in a reasonabble period of time which
leaves approximation as a better and desired alternative, and this is
where greedy algorithms do help. An example is the travelling salesman
problem, it takes O(n!) to find the best answer while an approximate
solution using greedy is O(n^2). This is mind blowing.

DP is hard but today I've tasted a gist of it. Grids, some type of smart
recursion. I need more understanding and examples.