Read on c++ utilities. Created a Pool and Semaphore
To understand the language c++, I have downloaded a book on the c++
standard library. It is specifically about c++17 standard, I chose 17 as
it seems a good place to start where there are no complexities on
modules, blah, blah, concepts. Just simple c++.
Using the `std::this_thread::yield`, I've created a simple semaphore and
tested it out by doing a whole bunch of tasks, simply sleeping, and I
noticed that, to do 100 tasks, I have to create 100 threads, this
doesn't scale very well as the number of tasks increases so I decided to
cook up a simple task processing pool. It tasks the number of workers to
use - number of threads to be spawned for processing tasks - which
receives a task generator - a function that returns a task object - in
which it loops asking for a new task once it finishes processing the
last one until working is set to false. The task generator is simply
popping from a queue of tasks or wait until one is added with put_task,
a method of pool object. This constructs are extremely buggy, I still
don't understand enough c++ for a clean implementaion, but they work
perfectly for illustrations.
I am actually impressed that I was able to do this, Like this is my
first pool ever, I have created a whole bunch of other synchronization
primitives before so, with enough tools, any can be created.
Also read on c++ smart pointers. Yeah, that's all.