A little more linear algebra mathematics.
Nothing amusing today except for creating a matrix that can rotate a 2d
vector in any angle, BOOM. Most of the work was in doing actual math on
paper and some few proofs. This simple matrix thingy can be used in 2d
games to rotate objects and moving them around as expected. GoodNight: