Tried to implement a thread pool in c++
This is something that I have done before but the last was just playing
with the idea but this, this is different. This implementation is meant
to be reusable but not production ready, still serves the purpose of
learning. It is also to explore my understanding of c++ types which has
proved difficult and I'm stopping here for today, I've sketched out a
the skeleton waiting to be filled with the good meat, type correctness.
The types part is why am trying it in c++ otherwise in python it would
have been complete by now, GoodNight.