Looked at python and c++ operators and a little OS
Read on Operating systems, a few pages of it were enough to day dream
about creating my own. The syscalls, scheduling and history of the topic
are mesmorizing and I hope to finish the book. The book is by Tanenbaum,
Operating systems design and implementation, it has a case study, minix,
and the whole book walks through implementing a whole operationg system
from chapter 1 to the last, 6, the best part of the book. They used c in
the code listings which is amazing given my wish to comprehend c.

More on binary operators, did you know that if both `__add__` and
`__radd__` return NotImplemented or are not defined, a TypeError is
raised that the operation '+' is not defined for type(left) and
type(right). Amazing.

For the expression `a == b`, `a.__eq__(b)` is called first, if it
returns any object other that NotImplemented, then that object is
returned as the result, the object is usually a bool, if it returns the
NotImplemented object, then `b.__eq__(a)` is called, if it also returns
NotImplemented, False is returned as the result of the equality check on
a and b. Neat, this is some deep stuff over here.