Another ThreadPool in python akin to an EventLoop
I am kind of obsessed with concurrent systems, like how is an operating
system implemented, how to languages implement generators, type systems,
async/await interfaces and lastly Fibers -- like coroutines and go
channels. I'm getting there though, today I played around with a simple
ThreadPool I cooked together using worker threads and Lock as a
synchronization primitive, amazing. I like the way it just takes form
from nowhere. Jumble up a few things here and there and wala! This one
is special though. Its not just a Thread Pool, it is also an EventPool,
sought of, doesn't work on events but it simply pops a task executes it
and does it again and again till stopped. Unlike my previous attempts,
this one shutsdown gracefully, a feature I had in mind from the start.
There is a book on game programming that I ought to read. Seems like a
good place to start to think about event driven systems, event
representation and keeping the application interactive while events are
processed in the background, I think this, this is the place to start,
it introduces the topic from a game perspective but really, the info can
be applied anywhere concurrency is needed and as a bonus, there are lots
of tips 'bout application design, unfortunately, I've tried downloading
the book but failed in all my attempts, hopefully I will find one.