Read on OS Shedulers and python's `__format__`
I think python's symplicity is the by product of compactful language
design, I mean, the designers of python hide the gory details in the
beautiful design of python and by using interfaces to activate the
different parts, example, dunder methods for operator overloading hide
the whole operator overloading python supports and amazingly, it doesn't
have to break the rules because it is a special feature of the language,
look at it, they chose a naming method that most people don't use to
their advantage: have you ever written a function with a name like
`__funcname__` in any language, yeah, I din't think so. I like this
approach unlike C++ and over languages that make some function names
special with worst of all,  the operators they are meant to overload,
but maybe the python's core devs saw this and learned from it, we will
never know. Otherwise, I've experimented a little bit here and there,
remember pickle, well, you can send python objects via the network by
dumping and loading on the client and server respectively and hence,
sent. In the process, I've looked at the `__format__` dunder method and
the so not used `format` function that uses the dunder method, does make
stringifying more fun, but as usual, I'm short of the tricks used to
efficiently use this feature. As for the OS, I finished the theory part
of processes and about to start a case study of the minix processes,
with that I implemented a small scheduler using the Lottery Scheduling
Algorithm, I'll be damned, it works like a charm. Though I believe there
are better ways of doing it, but this, this is my first scheduling
algorithm, as usual, a whole bunch of dead ideas back to life. For the
rest of the changed files, they are just a result of over work and
moving around some stuff, NIGHT.