ByteOrder and test python's sum function
ByteOrder, closely related to endianness. I've not understood the terms
and reasonning that go into this topic of byte representation in memory,
but I've done enough reading to undestand text about it. First, there
are two mostly used endians used, little and big endian. Another one is
PDP endian. Big endian is the most relatable and easily understandable,
the most significant byte goes first followed by the next most
significant byte all the way to the least significant byte while little
endian the least significant byte goes first then the next least
significant all the way to the most significant byte. To put this in
english, consider the number 16961, which in hexadecimal is 0x4241, you
can clearly see it is a 16bit integer made of two bytes, 42 and 41, they
translate to letters `B` and `A` respectively. In big endian the memory
layout is as the number representation we can see above, 4241 thats it,
and this is why big endian is relatable while little endian goes 4142.
By the way, the most significant byte in that number, 0x4241, is 0x4200
and the least significant is 0x0041. The significance here is, to my
understanding, which byte best explains the whole, or is closest to the
whole, example 0x4200 is 16896 in decimal while 0x0041 is 65 hence
0x4200 is more significant than 0x0041.
For the python's sum part, there is some kind of warning in trying to
get the sum of a non-numeric iterable using sum which I tested to see if
it raises any kind of error, I don't know where I saw this but sum was
kind of designed best for numeric iterables.
Before I go, check out the byteorder.cpp file, it is full of other cool
stuff I learned from `languages` project. You can split a big number
into bytes then store them as you wish and recreate it by or'ing and
shifting the bytes to get the original one back, maybe this is the
secret of pythons massive integers. The file is just a master piece from
the beautiful use of preprocessing to just the beautiful comments and
sections that describe themselves, its a work of art. NIGHT.