Arg evaluation order, unreachable and thread local
These are some of the newly learned concepts in c/c++ and python.
Argument evaluation order is how the arguments to a function call are
evaluated, from first to last. This is implementation defined according
to c/c++ compilers. For clang and clang++ the evaluatio order is from
first to last argument expressions while gcc and g++ are in the reverse
order. It is facsinating that they even considered this pat of the
compiler, you would think they would just use first to last and be done
with it. My bservation of these tools -- llvm, g++/gcc, clang/clang++,
python -- is that their power comes from defining as much of the syntax
as possible without adding anymore. Instead of a new keyword from some
feature, they use an already exesting keyword, define some new context
around it and wala, the feature is up and running, a general rule is
that each new feature should eliminate atleast two problems/be some new
solutions according to Bjarne. For thread locals, well, they are
different per thread and therefore each thread may have a different
value sttached to a thread local since they are different per thread.
That's all folks, till next time.