A little more meta, folding and reviews.
My first code review, I do really write bad poor unmaintainable code I
should say. I've gone through some of my old projects like movie-app
first version I buoy, is it ugly. I'm ashamed to say I wrote it but it
is what it is. I wouldn't say movie-app2, the second better version, is
any better but sure bits the first by a tiny bit. I have to master the
art of abstraction, it seems super powerful and from my perspective it
is the ultimate weapon in code design, the ability to dumb it down to
only the necessary properties of an object so that it can interact with
everything that has the properties, tell me that is not nothing but
perfection. I think the first step is reasonning in protocols and
policies but I sure don't know what those are but I've grown, my code
now is better not much though. I have a better grasp of separation of
concerns than I did before, thanks to the few design patterns I've been
looking at, though I'm not sure that is enough. I believe in practice,
the way I get better at this is making projects, thinking hard about the
design decisions I make and try to foresee their effects. But at the
same time, I don't wanna bore this hobby of mine with endless thinking.
I'm sticking to the old ways, create projects then review them to see
what can be improved, I can't say this is better as I can waste lots of
time on work that would have taken way shorter time if I had just
thought before the starting. Whoooooo, is it hot in here.
Otherwise, I wanted to see how unfolding works, I don't wanna say much,
just go to the necessary files. For the meta part, I've found a book on
python metaprogramming, skimmed it and it looks promissing. More on it
later, for now, NIGHT.