Dataapp Review and lots of math
Took a random project, dataapp, created using aiohttp and reviewed it.
To say the truth, it is the one project in this repo that is bearable
given my current capability. I think this is one of the projects I made
with the aim of understanding and using bit masks. The masks represented
permission levels, etc, readon the project code for comprehension.
Otherwise, I've done lots of math today, none involved programming,
school work. As I did all that, an idea was born, like the last one, it
seems pretty important and best kept as a secret. I've thought about it
for a significant amount of time and here I am adding it to my diary; after this. NIGHT.