Read and experimented with IEEE754 double spec
Not much, but '0/0' in C does not make the program crash, it returns a
double that means NaN according to IEEE754 spec, this is a double with
its first bit being 0, the following 11 being 1 and the last 52 can be
any permutations of one's and zero's. This in total makes up 64 bits
which is the size of the double. The important part is the 11 bits, they
are the defining property of the NaN. According to my machine, if the
sign bit is on then the double is negative and if false it is positive,
if the next 11 bits are on and the rest, 13th bit is off the and the
rest be in any permutation then they represent infinity; positive or
negativity is again determined by the sign bit. If the 13th bit is on in
the above sequence then it represents nan which again has positivity and
negativity represented by the sign bit. This is totally awesome.
And also python has the nan concept, `float('inf') - float('inf')` and
`float('nan'` will return the NaN float. And also, nan cannot equal
itself even if they are made of the same exact sequence of 1's and 0's,
but you can still compare equality if you view that sequence as a 64 bit
integer and compare which will return true.

Along the way I created a bit printer, it is capable of printing bits
that make up a number, some clever bit mangling was involved and it was
fun given that I did it in C. That's all my good people.