Refactored a little
The movie-app-2 needs a complete rewrite, there is a lot of coupling,
changing/adding anything breaks the whole thing. Otherwise, I was able
to add a few sorters as well as put them all in a single module which is
a better module compared to staticmethods in classes; God knows what I
was thinking when designing that. Night.