movie-app-2 broken from aiohttp version bump
The middleware feature to redirect missing resources to default ones
broke. The main problem being how aiohttp now handles stream response,
the middleware exists without catching NotFound error since the
FileResponse is Streamed, meaning the chunking is orchastrated by
aiohttp in the background after the middlewares exit. To fix this, we
link the missing resources to the default ones in the file system and
allow symbolic_links to resolve. Perfect!