Thread Pool in C using <threads.h>
Okay, there's a first time for everything, and this is the first for
multiple things. One, I finally used a conditional variable, last time I
wrote a thread pool was in C++ and I struggled in keeping the workers
busy when there were no tasks, I ended up using yield, works but really
innefficient as that consumes precious cycles that could be doing
something else, this is busy waiting while condition variables make the
thread go to sleep until it is woken up, cool. Two, to make the task
queue more efficient, I allocate slabs of memory to prevent allocating
small objects only to then free them, this is kinda cool, the queue is a
linked list of these nodes making it a bit faster. And finally, three,
my firt async C code, or perhaps, first weird program, the errors for
this one were just frustrating to catch, resulting in weird debugging,
unfortunately, after all I've gone through, I still used thread
yielding, as it turns out, a thread that locks must be the same that
unlocks, making mutexes obsolete for pool waiting, the current strategy
works but can't tell for sure if it is the best, but hey, it works.
This can be easily ported to pthreads but that'snot my main problem,
I've used pthreads before, so no worry, another header in my toolbelt
still.