Big Numbers in C++, speed test on fib generation: PY vs C++
Using the GMP library wrapped by GMPXX for c++, I was able to make C++
compete with python in fibonacci number generation. It computes the
first 60000 fibonacci numbers storing them in a file `fibs.(py|cpp).txt`
The numbers are so big that we get files of size 359M each and the last
number having over 12000 digits, wow! To achieve this,we needed GMP and
enbling big numbers in python, atleast the conversion of big ints to str
using `sys.set_int_max_str_digits(0)`, the default is 4300 which
wouldn't suffice.